Rename Readme to Readme.md
Gender-Classification-using-Scikit-
This is a code for gender classification. The code uses the scikit-learn machine learning library to train a decision tree on a small dataset of body metrics (height, width, and shoe size) labelled male or female. Then we can predict the gender of someone given a novel set of body metrics.

Dependencies
•	Scikit-learn (pip install -U scikit-learn ) •	numpy (pip install numpy) •	scipy (pip install scipy)

I am using 4 classifiers from scikit to train on the same dataset and comparing their results. Accuracy is also determined by trying to predict testing I trained classifier on samples from the training data and see if it correctly classifies it.